refactor node fields for better maintainability
